refactor: extract song matching functionality into its own file
